+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.3.2] - 2025-07-18
+### Added
+- Add checkout type to order custom fields
+
+### Changed
+- Consolidate additional_info items to single item when is Mercado Pago saved card flow
+
 ## [8.3.1] - 2025-07-16
 ### Changed
 - Do not display taxes in Argentina when data is not obtained